+++ src/Framework/Components/AnimatedSprite.h
@@ -13,7 +13,7 @@ namespace nge {
                 SDL_Rect dst,
                 int frameCount,
                 int repeats = -1,
-                double frameDuration = 1,
+                int frameDuration = 1,
                 double angle = 0.0f,
                 SDL_Point rotationCentre = ROTATION_CENTRE,
                 SDL_RendererFlip flip = SDL_FLIP_NONE
